+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
-### vNEXT
+### Next Version
+* Basic fullscreen support for Android MediaPlayer [#1138](https://github.com/react-native-community/react-native-video/pull/1138)
 * Support video cachging for iOS ([#955](https://github.com/react-native-community/react-native-video/pull/955))
 
 ### Version 3.1.0